+++ plugins/voxelman/eventdispatcher/plugin.d
@@ -3,24 +3,7 @@ module voxelman.eventdispatcher.plugin;
 import pluginlib;
 import voxelman.core.config;
 
-abstract class GameEvent {
-}
-
-template isGameEvent(T)
-{
-	enum bool isGameEvent = is(typeof(
-    (inout int = 0)
-    {
-    	T t;
-    }));
-}
-
-private class ValidGameEventClass : GameEvent {}
-private struct ValidGameEvent {}
-static assert(isGameEvent!ValidGameEvent);
-static assert(isGameEvent!ValidGameEventClass);
-
-private alias EventHandler = void delegate(ref GameEvent event);
+private alias EventHandler = void delegate(ref Object event);
 
 shared static this()
 {
@@ -34,13 +17,11 @@ class EventDispatcherPlugin : IPlugin
 
 	void subscribeToEvent(Event)(void delegate(ref Event event) handler)
 	{
-		static assert(isGameEvent!Event);
 		_eventHandlers[typeid(Event)] ~= cast(EventHandler)handler;
 	}
 
 	void postEvent(Event)(auto ref Event event)
 	{
-		static assert(isGameEvent!Event);
 		auto handlers = typeid(Event) in _eventHandlers;
 		if (!handlers) return;